feat: Configure Multer and S3 file uploads to bucket location.
