+++ server/app.js
@@ -2,12 +2,11 @@ const express = require('express')
 const dotenv = require('dotenv')
 const cors = require('cors')
 const multer = require('multer')
-const fs = require('fs')
 
 const connectDatabase = require('./config/database.config')
+const { storage, imageFilter } = require('./config/multer.config')
+const { s3Upload } = require('./config/aws.config')
 const userRouter = require('./routes/user.routes')
-const s3 = require('./config/aws.config')
-
 
 const app = express()
 dotenv.config()
@@ -19,28 +18,11 @@ app.use(express.json())
 
 app.use(userRouter)
 
-const storage = multer.diskStorage({
-  filename: (req, file, callback) =>
-    callback(null, `${Date.now()}-${file.originalname}`)
-})
-
-const imageFilter = (req, file, callback) => {
-  if (!file.originalname.match(/\.(jpg|jpeg|png|gif)$/i))
-    return callback(new Error('Only image files are allowed!'), false)
-  callback(null, true);
-}
-
 app.post('/api/image-upload', multer({ storage, fileFilter: imageFilter }).single('image'), async (req, res, next) => {
   try {
-    const DIRECTORY = '/test'
-    
-    const uploadedImage = await s3.upload({
-      Bucket: process.env.S3_BUCKET_NAME + DIRECTORY,
-      Body: fs.createReadStream(req.file.path),
-      Key: `${Date.now()}-${file.originalname}`
-    }).promise()
-  
-    console.log(uploadedImage)
+    const imageLocation = await s3Upload(req.file.originalname, req.file.path, req.body.directory)
+    console.log(imageLocation)
+
     return res.status(200).json({ type: 'success', msg: 'Image was uploaded successfully.' })
   } catch (error) {
     console.error(error)
